# docs/swagger.yaml
basePath: /
definitions:
...
  domain.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  domain.Heartbeat:
    properties:
      duration:
        type: integer
      log_id:
        type: integer
    type: object
  domain.HubService:
    properties:
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      last_update:
        type: string
      name:
        type: string
      price:
        type: number
      screenshot_url:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
    type: object
...
  domain.PublicService:
    properties:
      app_url:
        type: string
      id:
        type: integer
      last_update:
        type: string
      name:
        type: string
      status:
        type: string
      version:
        type: string
    type: object
...
  domain.Service:
    type: object
  domain.SuccessResponse:
    properties:
      data:
        description: Omits Data field if nil
      message:
        type: string
    type: object
  domain.UseService:
    properties:
      log_id:
        type: integer
      service:
        $ref: '#/definitions/domain.PublicService'
    type: object
...
host: 127.0.0.1:8080
info:
  contact:
    email: suporte@solude.tech
    name: Eng. Gabriel Coelho | InovaData development team
    url: https://solude.tech
  description: Platform API is a RESTful API for managing ...
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Platform API
  version: 0.1.1
paths:
...
  /services:
    get:
      description: Gets all available services
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              allOf:
              - $ref: '#/definitions/domain.SuccessResponse'
              - properties:
                  data:
                    items:
                      $ref: '#/definitions/domain.PublicService'
                    type: array
                type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Fetch Services
      tags:
      - Service
    post:
      consumes:
      - application/json
      description: Creates a new service
      parameters:
      - description: Service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/domain.Service'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.PublicService'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Create Service
      tags:
      - Service
  /services/{identifier}:
    get:
      description: Gets service by numeric ID (e.g., /services/123) or name (/services/my-service)
      parameters:
      - description: Service ID or Name
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.PublicService'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get Service by Identifier
      tags:
      - Service
  /services/{serviceID}:
    delete:
      description: Deletes a service
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Delete Service
      tags:
      - Service
    put:
      consumes:
      - application/json
      description: Updates service data
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: integer
      - description: Service data
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/domain.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.PublicService'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Update Service
      tags:
      - Service
  /services/{serviceID}/application:
    get:
      consumes:
      - application/json
      description: Logs that a user started using a service, returns log ID and public
        service data
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UseService'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Start using a service (create a usage log)
      tags:
      - Service
  /services/{serviceID}/organization/{organizationID}:
    post:
      description: Links the service to an organization
      parameters:
      - description: Service ID
        in: path
        name: serviceID
        required: true
        type: integer
      - description: Organization ID
        in: path
        name: organizationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Set Service Availability
      tags:
      - Service
  /services/heartbeat:
    patch:
      consumes:
      - application/json
      description: Adds usage duration (in seconds) to a log record
      parameters:
      - description: Heartbeat data
        in: body
        name: heartbeat
        required: true
        schema:
          $ref: '#/definitions/domain.Heartbeat'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Heartbeat usage
      tags:
      - Service
...
  /services/organization:
    get:
      description: Gets all services linked to an organization
      parameters:
      - description: Organization ID
        in: path
        name: organizationID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/domain.HubService'
              type: array
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Get Services by Organization
      tags:
      - Service
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"